New Update to Outfits
At the Moment, I have a working DB for Outfits, with a ForeignKey and an @Ignore statement to prevent the constructor from adding the list to the DB. Working on the UI to get to the point to add at least Tops and Bottoms but I need a break for now.

Please let me know if you have questions.