New Update for AddOutfit
Theres A lot going on so far, and still a lot to add.

Clothing, Outfit, and OutfitWithClothing are the new Entity Classes so you will need tore-install the app to make the DB changes.

- using AutoGenerate was a bad practice, and with a little time, I was able to fix that situation. Clothing and Outfit id's are now null, and SQLite will automatically generate a Long number for us.
- The Outfit Class is not working as of yet, only because it is going to take time to implement each Recyclerview. As of now, I feel confident in how to approach this solution as well as future solutions using the new setup.
- GSON, was a terrible approach, and honestly I'm glad we aren't using it.

** IMPORTANT **
I was told based on our practice for our application, its better to keep everything in one DAO, Repository, and ViewModel class. So all the data going form the Database to the ViewModel are renamed:
ClosetDao, ClosetRepository, and ClosetViewModel. Makes our lives easier only referencing the ClosetViewModel when working in the UI classes.